+++ library/src/schemas/object/objectAsync.test-d.ts
@@ -1,5 +1,5 @@
 import { describe, expectTypeOf, test } from 'vitest';
-import type { ReadonlyAction } from '../../actions/index.ts';
+import type { ReadonlyAction, TransformAction } from '../../actions/index.ts';
 import type { SchemaWithPipe } from '../../methods/index.ts';
 import type { InferInput, InferIssue, InferOutput } from '../../types/index.ts';
 import type { NullishSchema } from '../nullish/index.ts';
@@ -47,6 +47,12 @@ describe('objectAsync', () => {
         key4: ObjectSchemaAsync<{ key: NumberSchema<undefined> }, never>;
         key5: SchemaWithPipe<[StringSchema<undefined>, ReadonlyAction<string>]>;
         key6: UndefinedableSchema<StringSchema<undefined>, 'bar'>;
+        key7: SchemaWithPipe<
+          [
+            OptionalSchema<StringSchema<undefined>, never>,
+            TransformAction<undefined | string, string>,
+          ]
+        >;
       },
       undefined
     >;
@@ -59,6 +65,7 @@ describe('objectAsync', () => {
         key4: { key: number };
         key5: string;
         key6: string | undefined;
+        key7?: string;
       }>();
     });
 
@@ -70,6 +77,7 @@ describe('objectAsync', () => {
         key4: { key: number };
         readonly key5: string;
         key6: string;
+        key7: string;
       }>();
     });