+++ library/src/schemas/objectWithRest/objectWithRestAsync.test-d.ts
@@ -1,5 +1,5 @@
 import { describe, expectTypeOf, test } from 'vitest';
-import type { ReadonlyAction } from '../../actions/index.ts';
+import type { ReadonlyAction, TransformAction } from '../../actions/index.ts';
 import type { SchemaWithPipe } from '../../methods/index.ts';
 import type { InferInput, InferIssue, InferOutput } from '../../types/index.ts';
 import type { BooleanIssue, BooleanSchema } from '../boolean/index.ts';
@@ -60,6 +60,12 @@ describe('objectWithRestAsync', () => {
         key4: ObjectSchema<{ key: NumberSchema<undefined> }, never>;
         key5: SchemaWithPipe<[StringSchema<undefined>, ReadonlyAction<string>]>;
         key6: UndefinedableSchema<StringSchema<undefined>, 'bar'>;
+        key7: SchemaWithPipe<
+          [
+            OptionalSchema<StringSchema<undefined>, never>,
+            TransformAction<undefined | string, string>,
+          ]
+        >;
       },
       BooleanSchema<undefined>,
       undefined
@@ -74,6 +80,7 @@ describe('objectWithRestAsync', () => {
           key4: { key: number };
           key5: string;
           key6: string | undefined;
+          key7?: string;
         } & { [key: string]: boolean }
       >();
     });
@@ -87,6 +94,7 @@ describe('objectWithRestAsync', () => {
           key4: { key: number };
           readonly key5: string;
           key6: string;
+          key7: string;
         } & { [key: string]: boolean }
       >();
     });